get httpclient from IHttpClientFactory
This is the basic way to use httpclientfactory, we could also get a little fancier and do a named "sloth client" or even make a strongly typed httpClient which just knows how to do sloth API calls.  This seems like a good middle-ground though for now.  Performance should be better and networking problems should be less.  

Closes #646