Copies history from change request project
When creating a project from a change request, copies the project history from the change request project to the new project. Only the PI visible ones

This allows users to see the history of the change request that led to the creation of the project.

The history entries are prefixed with "From Change Request [ID]:" to indicate their origin.