Adds links to permission details
Improves user role management by adding links to the
details view for worker and supervisor permissions on the
permissions index page. This allows administrators to
easily view and manage individual permissions assigned
to users within a team. A details link is also added to
the details page for a specific worker.