Adds mobile API endpoint for projects
Adds an API endpoint to retrieve a list of active projects for a given team.
This endpoint is intended for use by mobile applications.

It authenticates the request using an API key passed in the Authorization header.
Only active projects associated with the team linked to the API key are returned.