Enables mobile app link without team param
Allows users to access the mobile app link without a team parameter in the URL.

If the team is not provided in the route, the application will attempt to
fetch the team from the API. If the user only belongs to one team, it's
automatically selected. If the user belongs to multiple teams or no team,
the user will be redirected to the team selection page.