Adds expense creation endpoint for mobile app
Adds an endpoint to allow mobile app users to create expenses for a project.

Validates the expenses to ensure that all expenses have the same WorkerMobileId, and that no expenses with the same WorkerMobileId already exist.
Also ensures that expenses can only be created for active projects and automatically approves expenses if the user has the FieldManager or Supervisor role.